+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Modules permissions (https://github.com/PierreRambaud/GotCms/issues/21)
 - Methods fetchAll, fetchOne, fetchRow and execute now have the ability to pass parameters
 - Backup module (https://github.com/PierreRambaud/GotCms/issues/20)
+- Update to Zend Framework 2.1
 
 ## 0.1.3 (12 Jan 2013):
 - Sitemap module QuickFix set base path (https://github.com/PierreRambaud/GotCms/issues/14)